tr1/objects/movable_block: add trap feature to pushblocks
This makes falling pushblocks more realistic if they land on top of
Lara, killing her outright. We hide Lara's shadow to avoid it
appearing on top of the pushblock, as the floor height change will
continue to operate in case there are active enemies nearby.
Additionally, a camera target item is spawned to prevent issues with
the camera looking into the pushblock at Lara.

Resolves #2035.