inventory: abandon backpack module
The backpack module was originally meant to focus distinctly on Lara's
item management and distinguish it from the inventory ring routines.
However, we decided it makes more sense to modularize the inventory ring
parts as InvRing, the inventory management as just Inv. Since the
backpack API was still in its early stages, with just two functions
around, this change should have little to no impact.