tr1/data: fix underwater lever animation
This adds a "hands-free" command at the end of Lara's underwater lever
animation, which is in line with TR2. It prevents issues after using an
underwater lever followed by entering the wading state. The additional
check in Lara's state control also matches TR2, which is one that we
missed before.

Resolves #2416.